# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CyCove Backend API
  version: 0.6.0-draft
  description: >
    HTTP API contract for the CyCove relay backend, formalizing the surface
    sketched in the vault note (Projects/CyCove.md, Architecture -> Backend).
    The server is a deliberate "dumb pipe": it never sees plaintext message
    content and never holds long-term encryption keys. Realtime delivery is
    covered separately in asyncapi.yaml.
    STATUS (2026-07-19): reshaped around matrix-sdk-crypto-wasm's actual
    protocol requirements — see docs/crypto-integration-notes.md for why.
    /keys/upload, /keys/query, /keys/claim replace the original /prekeys
    endpoints (Matrix-shaped batch operations, not this API's own simpler
    shapes). user_id and device_id are now client-generated, not
    server-assigned — required because the client's OlmMachine bakes both
    into a self-signed device_keys blob before /register is ever called.
    STATUS (2026-07-19, later same day): added a real username directory
    (/users/*), rate limiting on the most abuse-prone endpoints, and opaque
    encrypted contact backup (/account-data/*) — see
    docs/crypto-integration-notes.md and THREAT_MODEL.md for the disclosures
    that come with both.
    STATUS (2026-07-20): added /login (re-authenticate an existing account on
    the device that created it) and wired the web client to the previously
    backend-only /devices/pairing-token and /devices/link (real "linked
...
    more use for it.
...
servers:
...
paths:
...
  /users/username:
    put:
      tags: [users]
      summary: Claim or change the caller's username
      description: >
        Reverses the original "local label only, no directory" identity
        model — see docs/crypto-integration-notes.md and
        Projects/CyCove.md -> Key decisions -> Identity model for why.
        Optional, unique, lowercase-only ASCII. Race-safe via the same
        unique-constraint-collision pattern as user_id/device_id elsewhere
        in this API.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                  pattern: '^[a-z0-9_]{3,32}$'
                  description: Normalized lowercase server-side regardless of input casing.
      responses:
        '200':
          description: Username claimed
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
        '400':
          description: Invalid format (must be 3-32 chars, lowercase letters/numbers/underscore only)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: username already taken
        '429':
          $ref: '#/components/responses/RateLimited'
  /users/by-username/{username}:
    get:
      tags: [users]
      summary: Resolve a username to an account and device
      description: >
        Case-insensitive exact match. Returns the account's
        most-recently-active device (Device.lastSeenAt) — a deliberate
        simplification given no real multi-device usage exists anywhere in
        the app yet; revisit once device linking is actually wired up
        client-side. This endpoint is the new discoverability surface — see
        THREAT_MODEL.md -> Server compromise for what a full DB dump now
        additionally exposes as a result.
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Match found
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  device_id:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No user with that username
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Device-scoped session token, issued by /register or /devices/link.
        Implemented as an opaque random token mapped to a device ID in Redis
        (backend/src/lib/auth.ts) — no expiry or refresh yet, that's still an
        open design question, not just an unimplemented one.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid session token
    RateLimited:
      description: >
        Rate limit exceeded. First-pass limits shipped 2026-07-19
        (backend/src/lib/rateLimit.ts, fixed-window, per-device or per-IP
        depending on the endpoint) — not yet tuned against real usage, see
        THREAT_MODEL.md -> Open items.